% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTraits_outputs.R
\name{BayesTraits.outputs}
\alias{BayesTraits.outputs}
\title{Process BayesTraits output data}
\usage{
BayesTraits.outputs(logst_dir = NULL,
                    responvar = NULL,
                    explanvar = NULL,
                    treetransf = c("Kappa", "Delta", "Lambda", "OU"),
                    bayesfactor = FALSE,
                    unirates = TRUE,
                    outformat = c("Word", "CSV"),
                    tableleg = "Phylogenetic regression in BayesTraits: models and coefficients.",
                    dir_create = "results_BayesTraits_output",
                    outfile = "BayesTraits_output_table",
                    ...)
}
\arguments{
\item{logst_dir}{Path to the folder directory where the Log and stepping
stone sampler files from the BayesTraits phylogenetic independent contrast
regression analyses are stored. The Log file(s) contain the model options and
output. The stepping stone sampler files file(s) contain the estimated log
marginal likelihood and the additional information about the stones. Make sure
that all these are also named with the explanatory variable(s) separated by
underscore, as well as the exact name of the tree transformation set in the analysis.
So, you have to name them either with \emph{Kappa}, \emph{Delta}, \emph{Lambda}, \emph{OU} (Ornstein–Uhlenbeck),
\emph{UNI} (if you have used an uniform tree transformation), \emph{VR} (if you have used
a variable rates) or \emph{Fabric} (if you have used the general statistical model
that accommodates uneven evolutionary landscape as described by \href{https://doi.org/10.1038/s41467-022-28595-z}{Pagel et al. (2022)}.
Thus, your Log file(s) should be name as something like _VAR1NAME_VAR2NAME.txt.Log.TREETRANSFORMATIONNAME.txt.
See the following real examples of Log input files: "BayesTraits_mean_data_bio12_bio15.txt.Log.VR"
and "BayesTraits_mean_data_bio12_bio15_nnodes.txt.Log.Lambda". This means that
the first Log file contains two models named "bio12" and "bio15" and was based
on VR (Variable Rates tree transformation), the second Log file contains three
models named "bio12", "bio15", and "nnodes", and was based on Lambda tree transformation.
Likewise, your Stones file(s) should be name as something like _VAR1NAME_VAR2NAME.txt.Stones.TREETRANSFORMATIONNAME.txt.
See the following real examples of Stones input files: "BayesTraits_mean_data_bio12_bio15.txt.Stones.VR"
and "BayesTraits_mean_data_bio12_bio15_nnodes.txt.Stones.Lambda". This means
that the first Stones file contains two models named "bio12" and "bio15" and
was based on VR (Variable Rates tree transformation), the second Stones file
contains three models named "bio12", "bio15", and "nnodes", and was based on
Lambda tree transformation.}

\item{responvar}{Give the name of the response variable.}

\item{explanvar}{Give the name(s) of the explanatory variable(s) exactly as
they are also written in the names of the Log and Stones files to be processed.}

\item{treetransf}{Select any of the available BayesTraits tree transformation
of evolution (\emph{Kappa}, \emph{Delta}, \emph{Lambda} or \emph{OU}), with which you have analysed
your data. If you have analysed your data with just Uniform, VR or Fabric models
then make sure you defined \code{treetransf = NULL}.}

\item{bayesfactor}{Logical, if \code{FALSE}, then no Bayes Factor pairwise comparisons
will performed among the models.}

\item{unirates}{Logical, \code{FALSE}, then the results with uniform tree transformation
will not be displayed in the main table.}

\item{outformat}{Define either "Word" or "CSV", or a vector c("Word", "CSV")
of both, for writing the results in any of such formats.}

\item{tableleg}{A legend for the main table, provided that you have also chosen
"Word" in the argument \code{outformat} for the resulting format. If you do not give
any specific legend here, then the Word-formatted table will have the default
legend \strong{Phylogenetic regression in BayesTraits: models and coefficients}.}

\item{dir_create}{Pathway to the computer's directory, where the file(s) will be saved.
The default setting creates a directory named \strong{results_BayesTraits_output} in
which the results will be saved within another subfolder named by the current date.}

\item{outfile}{Name of the main resulting table-formatted files in either "Word"
or "CSV" (or both formats; i.e. depending on the specified argument \code{outformat})
to be saved on the directory. If no name is given, the default setting creates
a file named \emph{BayesTraits_output_table.docx} and/or \emph{BayesTraits_output_table.csv}.}

\item{...}{additional parameters passed to pdf.}
}
\value{
Table in dataframe which is also saved on disk in .csv format.
}
\description{
It processes the output Log and Stones .txt files from phylogenetic
regression analyses with Meade & Pagel's (2022) \href{http://www.evolution.reading.ac.uk/BayesTraitsV4.0.0/BayesTraitsV4.0.0.html}{BayesTraits}
computer package. The resulting processed files are tables either in Word
or CSV formats summarizing each model, including comparison with Bayes Factor.
Heatmaps based on pairwise Bayes Factor comparison are also automatically produced
in PDF format. See \href{http://www.evolution.reading.ac.uk/BayesTraitsV4.0.0/Files/BayesTraitsV4.0.0-Manual.pdf}{BayesTraits V4.0.0 Manual}
for further details on each of the output files produced by BayesTraits so as to better
understand the processed output files from the current function \code{BayesTraits.outputs}.
}
\seealso{
\code{\link{BayesTraits.inputs}}
}
\author{
Domingos Cardoso & Matt Lavin
}
