% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTraits_shell.R
\name{BayesTraits.shell}
\alias{BayesTraits.shell}
\title{Create shell scripts to run BayesTraits phylogenetic regression analyses}
\usage{
BayesTraits.shell(meanfile_dir = NULL,
                  linkfile_dir = NULL,
                  treefile_dir = NULL,
                  BayesTraits_dir = NULL,
                  responvar = NULL,
                  treetransf = NULL,
                  bi = "30000000",
                  it = "130000000",
                  sa = "100000",
                  st = c("500", "50000"),
                  syst = NULL,
                  dir_create = "results_BayesTraits_shell",
                  cc_DataTree = TRUE)
}
\arguments{
\item{meanfile_dir}{Path to the folder directory where the mean data file(s)
is(are) stored.}

\item{linkfile_dir}{Path to the folder directory where the linked data file(s)
is(are) stored.}

\item{treefile_dir}{Path to the folder directory where the tree file is stored.}

\item{BayesTraits_dir}{Path to the folder directory where the BayesTraits
executable file is stored.}

\item{responvar}{Give the name of the response variable of your analysis.}

\item{treetransf}{Define any or all (listed as a vector) available tree transformation
models that you want to run BayesTraits phylogenetic independent contrast regression
analyses with your input data. As such, choose among the models \emph{Kappa}, \emph{Delta},
\emph{Lambda}, \emph{OU} (Ornstein–Uhlenbeck), \emph{UNI} (if you have used an uniform tree
transformation), \emph{VR} (if you have used a variable rates) or \emph{Fabric} (if you
want to use the general statistical mode that accommodates uneven evolutionary
landscape as described by \href{https://doi.org/10.1038/s41467-022-28595-z}{Pagel et al. (2022)}.}

\item{bi}{Set the number of iterations to burn the MCMC chain in for. Please
give the number as a character, i.e. under quotes.}

\item{it}{Set the number of iterations to run the chain for. Please give the
number as a character, i.e. under quotes.}

\item{sa}{Set the sample frequency. Please give the number as a character,
i.e. under quotes.}

\item{st}{The stepping stone sampler to estimate the marginal likelihood. Set
the number of stones to use and the number of iterations to use each stone for.
Please give the numbers as characters, i.e. a vector of numbers under quotes.}

\item{syst}{Give the name of the response variable.}

\item{dir_create}{Pathway to the computer's working directory, where the file(s)
will be saved. The default setting creates a directory named "results_BayesTraits_output" in
which the results will be saved within another subfolder named by the current date.}

\item{cc_DataTree}{Logical, the default is \code{TRUE}, then a copy of the tree,
mean and linked data files will be made directly into the BayesTraits folder.}
}
\description{
Based on the directory paths to the tree, linked and mean data files,
it automatically produces shell scripts in both Unix and Windows PowerShell
formats for running one or multiple phylogenetic regression analyses with
Meade & Pagel's (2022) \href{http://www.evolution.reading.ac.uk/BayesTraitsV4.0.0/BayesTraitsV4.0.0.html}{BayesTraits}
computer package. In addition to generate the shell script directly in BayesTraits
directory, the function also creates all running files with BayesTraits specified
parameters (bi, it, sa, and st) for any run, depending on the chosen model of
tree transformation. See \href{http://www.evolution.reading.ac.uk/BayesTraitsV4.0.0/Files/BayesTraitsV4.0.0-Manual.pdf}{BayesTraits V4.0.0 Manual}
for further details on each of BayesTraits parameters available for phylogenetic
regression so as to better understand how to set the arguments of the current
function \code{BayesTraits.shell}. You might also want to look at how to use
the function \code{\link{BayesTraits.inputs}} to produce automatically all linked
and mean data input files prior to using \code{BayesTraits.shell}.
}
\seealso{
\code{\link{BayesTraits.inputs}}

\code{\link{BayesTraits.outputs}}
}
\author{
Domingos Cardoso & Matt Lavin
}
