% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTraits_shell_run_proj.R
\name{BayesTraits.shell.reg.proj}
\alias{BayesTraits.shell.reg.proj}
\title{Create a BayesTraits project folder with associated example scripts and input
data for a full shell script run and log processing of BayesTraits phylogenetic
regression analyses}
\usage{
BayesTraits.shell.reg.proj(dir_create_proj = NULL,
                           syst = NULL)
}
\arguments{
\item{dir_create_proj}{Path to any computer's directory where an example
of a BayesTraits shell project folder and associated files for a full run with
shell script should be saved.}

\item{syst}{Report the operating system ("unix" or "windows").}
}
\description{
By reporting a directory path and the computer's operating system
("unix" or "windows"), this function creates a BayesTraits project folder with
associated example scripts and input data for a full shell script run and log
processing multiple phylogenetic regression analyses with Meade & Pagel's
(2022) \href{http://www.evolution.reading.ac.uk/BayesTraitsV4.0.0/BayesTraitsV4.0.0.html}{BayesTraits}
program. The function produces an example of a shell script in Unix or Windows PowerShell
that helps to fully automatize both the complete BayesTraits regression analysis
and processing of the resulting log and stones files, by automatically running
the three main \code{InNOutBT} functions \code{\link{BayesTraits.inputs}},
\code{\link{BayesTraits.shell}}, and \code{\link{BayesTraits.outputs}}, without
directly opening the R or R Studio programs. Note that these files are just
examples from an specific study of the Vataireoid legumes, and so you will have
to edit the folder paths and analysis parameters in such example scripts according
to your desired regression analyses. See also a more complete article
\href{https://dboslab.github.io/InNOutBT/articles/independent_contrast_regression_shell_project.html}{describing a BayesTraits shell run project}
that can be created with \code{BayesTraits.shell.reg.proj} for fully automated
multiple phylogenetic regression analyses.
}
\examples{
\dontrun{
library(InNOutBT)

BayesTraits.shell.reg.proj(dir_create_proj = "/Users/domingoscardoso/Downloads",
                           syst = "unix")
}

}
\seealso{
\code{\link{BayesTraits.inputs}}

\code{\link{BayesTraits.shell}}

\code{\link{BayesTraits.outputs}}
}
\author{
Domingos Cardoso & Matt Lavin
}
