% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vatsTree}
\alias{vatsTree}
\title{BEAST-dated phylogeny or time-tree of the Vataireoid legumes}
\format{
Treedata file in nexus format as resulting from BEAST analysis
}
\source{
\href{https://doi.org/10.12705/675.6}{Cruz et al. (2018)}
}
\usage{
data(vatsTree)
}
\description{
A dated phylogeny of the Vataireoid legumes from a BEAST
analysis of nine genes (nuclear ribosomal ETS and ITS, and the plastid regions
matK, 3′-trnK intron, psbA-trnH, trnL intron, rps16 intron, trnQ-5′rps16,
and trnD-trnT).
}
